set failing code for url-loader, if data fetching fails
