+++ url-loader/src/database/attributes.js
@@ -8,14 +8,14 @@ const setupCollectionForAttributes = () => {
   if (config.mongoUri) {
     return mongoDb.createCollection(ATTRIBUTES_COLLECTION_NAME, { date: -1 });
   }
-  throw new Error('No connection string to database');
+  throw new Error('No credentials for database');
 };
 
 const getLatestAttributeDate = (attributeId) => {
   if (config.mongoUri) {
     return mongoDbAttributes.getLatestAttributeDate(attributeId);
   }
-  throw new Error('No connection string to database');
+  throw new Error('No credentials for database');
 };
 
 module.exports = { setupCollectionForAttributes, getLatestAttributeDate };
