+++ url-loader/src/database/index.js
@@ -9,21 +9,21 @@ const initializeDb = () => {
   if (config.mongoUri) {
     return mongoDb.initializeDBConnection();
   }
-  throw new Error('No connection string to database');
+  throw new Error('No credentials for database');
 };
 
 const disconnectFromDB = () => {
   if (config.mongoUri) {
     return mongoDb.disconnectFromDB();
   }
-  throw new Error('No connection string to database');
+  throw new Error('No credentials for database');
 };
 
 const createCollection = (collectionName, index, geoIndex) => {
   if (config.mongoUri) {
     return mongoDb.createCollection(collectionName(collectionName, index, geoIndex));
   }
-  throw new Error('No connection string to database');
+  throw new Error('No credentials for database');
 };
 
 module.exports = { initializeDb, disconnectFromDB, createCollection };
